Split code generators among multiple files
